requirements.txt
- asgiref==3.6.0
- crispy-tailwind==0.5.0
Django==4.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
Pillow==9.5.0
sqlparse==0.4.3
tailwind==3.1.5b0
